quickdrop implemented, fixing bugs related to interactions with placed blocks in progress
