+++ src/App.tsx
@@ -3,6 +3,7 @@ import './App.css'
 import styled from 'styled-components'
 import { GameCanvas } from './GameCanvas'
 import { blockControl } from './blockControl'
+import { setCanvas } from './setCanvas'
 
 const ContainerDiv = styled.div `
 display: flex;
@@ -64,7 +65,7 @@ function App() {
 
   const canvasRef = useRef<HTMLCanvasElement>(null);
 
-  const shape = useRef("I");
+  const shape = useRef("T");
   const rotation = useRef(1);
   const centerPoint = useRef([5, 5]);
 
@@ -75,8 +76,8 @@ function App() {
   }, []);
   
   useEffect(() => {
-    console.log({gridArr});
-    const canvas = canvasRef.current;
+
+    setCanvas(gridArr, gameDimensions, canvasRef);
     
     const handleKeyDown = (event: KeyboardEvent) => {
       console.log( event.key );
@@ -85,27 +86,6 @@ function App() {
     };
 
     document.addEventListener('keydown', handleKeyDown);
-
-    if (canvas) {
-      const ctx = canvas.getContext('2d');
-      if(ctx) {
-        gridArr.forEach((row, yInd) => {
-          row.forEach((item, xInd) => {
-            if(item === "[x]") {
-              ctx.fillStyle = "#7cce70ff"
-            } else if(item === "") {
-              ctx.fillStyle = "#222222ff"
-            } else {
-              ctx.fillStyle = "#47cc33ff"
-            }
-            ctx.fillRect(xInd*(gameDimensions[0]/10), yInd*(gameDimensions[1]/20), gameDimensions[0]/10, gameDimensions[1]/10);
-            ctx.rect(xInd*(gameDimensions[0]/10), yInd*(gameDimensions[1]/20), gameDimensions[0]/10, gameDimensions[1]/10);
-          });
-        });
-        ctx.strokeStyle = "#aaaaaaff";
-        ctx.stroke();
-      }
-    }
     
     return () => {
       document.removeEventListener('keydown', handleKeyDown);
